+++ packages/opentelemetry-resources/src/detectors/platform/node/OSDetector.ts
@@ -14,9 +14,19 @@
  * limitations under the License.
  */
 
-import { noopDetectorSync } from '../../detectors/NoopDetectorSync';
+import { Detector } from '../../../types';
+import { ResourceDetectionConfig } from '../../../config';
+import { IResource } from '../../../IResource';
+import { osDetectorSync } from './OSDetectorSync';
 
 /**
- * @experimental
+ * OSDetector detects the resources related to the operating system (OS) on
+ * which the process represented by this resource is running.
  */
-export const serviceInstanceIdDetectorSync = noopDetectorSync;
+class OSDetector implements Detector {
+  detect(_config?: ResourceDetectionConfig): Promise<IResource> {
+    return Promise.resolve(osDetectorSync.detect(_config));
+  }
+}
+
+export const osDetector = new OSDetector();